correcao de filtro pessoa
foi corrigido o local da implementacao de filtro de pessoa
